Refactoring build scripts
Making build target a build parameter instead of an ENV. Also creating lists to make it easier to read the build script. Removing from build.bat the test build and install and moving it to run_tests.bat, which now builds the game, builds the tests and run tests. run_tests.bat also receive the target as a parameter (release or debug)