Creating test source exclusive for setups and teardowns
Setups and teardown functions are highly reusable. For this reason i'm isolating them in a source file in order to use the same setup and teardown several times for several test functions. Also creating a macro that allow a setup function call another one that do the same it needs (and the only difference is the teardown)