Adding to test script the hability to check if build target was passed
It's necessary in order to avoid test script calling the build script of the game with "test" as the only argument, which does not enable the test build option because the build script has positional parameters, and test must be the second one. If test script receive no argument, when it tries to pass its first parameter as argument to build script, it expands to nothing and test becomes the single argument passed, which lead to errors. Test build does not work properly if game build does not have test enabled because the static symbols wont become public